- Added role based access control - Reverted login limit (to be added back in future commit)

+++ lottery/views.py
@@ -4,7 +4,7 @@ import logging
 from flask import Blueprint, render_template, request, flash
 from flask_login import login_required, current_user
 from sqlalchemy import desc
-from app import db
+from app import db, requires_roles
 from models import Draw, User
 
 # CONFIG
@@ -18,11 +18,14 @@ Adraw_key = Auser.draw_key
 # view lottery page
 @lottery_blueprint.route('/lottery')
 @login_required
+@requires_roles('user')
 def lottery():
     return render_template('lottery.html')
 
 
 @lottery_blueprint.route('/add_draw', methods=['POST'])
+@login_required
+@requires_roles('user')
 def add_draw():
     submitted_draw = ''
     for i in range(6):
@@ -44,6 +47,8 @@ def add_draw():
 
 # view all draws that have not been played
 @lottery_blueprint.route('/view_draws', methods=['POST'])
+@login_required
+@requires_roles('user')
 def view_draws():
     # get all draws that have not been played [played=0]
     playable_draws = Draw.query.filter_by(played=False, user_id=current_user.id).all()# TODO: filter playable draws for current user
@@ -68,9 +73,11 @@ def view_draws():
 
 # view lottery results
 @lottery_blueprint.route('/check_draws', methods=['POST'])
+@login_required
+@requires_roles('user')
 def check_draws():
     # get played draws
-    played_draws = Draw.query.filter_by(played=True).all()  # TODO: filter played draws for current user
+    played_draws = Draw.query.filter_by(played=True, user_id=current_user.id).all()  # TODO: filter played draws for current user
 
     played_draw_copies = list(map(lambda x: copy.deepcopy(x), played_draws))
 
@@ -92,8 +99,10 @@ def check_draws():
 
 # delete all played draws
 @lottery_blueprint.route('/play_again', methods=['POST'])
+@login_required
+@requires_roles('user')
 def play_again():
-    delete_played = Draw.__table__.delete().where(Draw.played)  # TODO: delete played draws for current user only
+    delete_played = Draw.__table__.delete().where(Draw.played, Draw.user_id == current_user.id)  # TODO: delete played draws for current user only
     db.session.execute(delete_played)
     db.session.commit()
